adjust SCSS & add Dummy data
1. adjust part of scss
2. refactor Todolist compoent to Layout folder
3. change name of "ListNav" to "AppNav"
3. add DUMMY Data, Path: public/dummydata.js
