+++ components/pages/AppPages/Project.js
@@ -1,75 +1,16 @@
 import { ListContentHeader } from "../../layout/ListContentHeader";
 import TodoList from "../TodoList";
-
-const DUMMY_DATA = [
-  {
-    id: 1,
-    content: "2learn React.js",
-    due: "2024-11-01",
-    project: "learn",
-    isChecked: false,
-    priority: "P1",
-  },
-  {
-    id: 2,
-    content: "4learn React.js",
-    due: "2021-01-01",
-    project: "learn",
-    projectId: 1000,
-    isChecked: false,
-    priority: "P1",
-  },
-  {
-    id: 3,
-    content: "2learn React.js",
-    due: "2024-01-01",
-    project: "learn",
-    description: "THis is what",
-
-    isChecked: false,
-    priority: "P1",
-    label: "Reading",
-  },
-  {
-    id: 4,
-    content: "6learn React.js",
-    due: "2024-01-01",
-    project: "learn",
-    description: "THis is what",
-
-    projectId: 1000,
-    isChecked: true,
-    priority: "P1",
-    label: "Reading",
-  },
-  {
-    id: 5,
-    content: "8learn React.js",
-    due: "2024-01-01",
-    description: "THis is what",
-    projectId: 1000,
-    isChecked: false,
-    priority: "P2",
-    label: "Reading",
-  },
-  {
-    id: 6,
-    content: "learn React.js",
-    due: "2024-01-01",
-    project: "learn",
-    projectId: 1000,
-    isChecked: false,
-    priority: "P1",
-    label: "Reading",
-  },
-];
+import { DUMMY_TASKS } from "../../../public/dummy-data";
 
 export default function Project({ projectId }) {
+  // fetch/filter relative tasks
+
+
   return (
     <>
       <ListContentHeader title={projectId} />
       <div className="list-box">
-        <TodoList datas={DUMMY_DATA} />
+        <TodoList datas={DUMMY_TASKS} />
       </div>
     </>
   );
