+++ styles/layout/TodoList.module.scss
@@ -1,7 +1,5 @@
 .list_editor {
   padding: 0px 55px 84px;
-
- 
 }
 
 .section {
