+++ styles/scss/utils/_basic.scss
@@ -81,3 +81,11 @@ i {
   background-color: $grayColor;
   color: $secondaryColor;
 }
+
+.click_close_cover {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+}
